+++ crates/moss-db/tests/integration__bincode_table__remove.rs
@@ -1,8 +1,9 @@
+mod shared;
+
 use moss_db::common::DatabaseError;
 use moss_db::DatabaseClient;
 
 use crate::shared::setup_test_bincode_table;
-mod shared;
 
 #[test]
 fn remove_success() {
